tests/integration/test_grpc_uds.py
"""
集成测试：测试 gRPC 通信链路（UDS）
"""

import pytest
import asyncio
import grpc
from pathlib import Path


@pytest.mark.integration
@pytest.mark.asyncio
async def test_uds_connection():
    """测试 UDS Socket 连接"""
    uds_path = "/tmp/test_cy_worker.sock"
    
    # 清理旧的 socket 文件
    if Path(uds_path).exists():
        Path(uds_path).unlink()
    
    # 这里应该启动一个测试用的 gRPC server
    # 由于需要实际的 server 实现，这里仅作为框架示例
    
    # 测试连接
    channel = grpc.aio.insecure_channel(f"unix://{uds_path}")
    
    # 清理
    await channel.close()
    
    # 简单断言（实际测试需要更多逻辑）
    assert True


@pytest.mark.integration
def test_coordinator_to_worker_flow():
    """测试 Coordinator -> Worker 完整流程"""
    # 这里应该：
    # 1. 启动 Worker (UDS)
    # 2. 启动 Coordinator
    # 3. 发送请求
    # 4. 验证响应
    
    # 由于需要完整环境，这里仅作为框架
    assert True
