pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "cy-llm-engine"
version = "0.1.0"
description = "CY-LLM Engine - Open Source Version with Rust Sidecar"
authors = [
    {name = "CY-LLM Team"}
]
readme = "README.md"
license = {text = "Apache-2.0"}
requires-python = ">=3.10"
dependencies = [
    # Configuration
    "pydantic>=2.0.0,<3.0.0",
    "pydantic-settings>=2.0.0,<3.0.0",
    
    # gRPC
    "grpcio>=1.60.0,<2.0.0",
    "grpcio-tools>=1.60.0,<2.0.0",
    "protobuf>=4.0.0,<6.0.0",
    
    # Utilities
    "numpy>=1.24.0,<2.0.0",
    "scipy>=1.10.0",
    "psutil>=5.9.0",
    "tqdm>=4.65.0",
    "xxhash>=3.0.0",
    "watchdog>=3.0.0",
    
    # HuggingFace ecosystem
    "transformers>=4.36.0,<5.0.0",
    "tokenizers>=0.15.0",
    "huggingface-hub>=0.20.0",
    "safetensors>=0.4.0",
    "datasets>=2.15.0",
    
    # Gateway
    "fastapi>=0.115.0",
    "uvicorn>=0.30.0",
]

[project.optional-dependencies]
nvidia = [
    "torch>=2.1.0",
]
vllm = [
    "vllm>=0.5.0",
]
tensorrt = [
    "tensorrt>=8.6.0",
]
training = [
    "peft>=0.7.0",
    "trl>=0.7.0",
    "sentencepiece>=0.2.0",
]
dev = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.21.0",
    "mypy>=1.7.0",
    "black>=23.0.0",
    "ruff>=0.1.0",
]

[tool.setuptools.packages.find]
where = ["src"]

[tool.black]
line-length = 100
target-version = ["py310", "py311"]

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.mypy]
python_version = "3.10"
strict = true
warn_return_any = true
warn_unused_configs = true
